[regalloc] Coalesce spill slots of phi-inputs and output eagerly.
Optionally pre-compute a coalescing of spill slots for values across
phis. The register allocator already did this on the fly when it
decided on spilling the output of a phi. This change makes that more
principled. Furthermore, as we now have a representation for
coalesced live ranges, also use that information to try assign
the same register on reloads.

Change-Id: I353c83ee3ef016bdf73e9d4c49cfeef58b9d8c7d