Update patch set 1
Patch Set 1: Commit-Queue+2

Relanding this, without any changes. It was not the cause of the failures. I've verified locally that:
(1) the test in question failed at the revision before the original commit as well
(2) with this reland applied to the current ToT, the test passes.

Patch-set: 1
Reviewer: Gerrit User 1115935 <1115935@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Commit-Queue=+2