Fix array spread on large sets and maps.
The fast paths for sets and maps did not support allocation
in large object space, yet they were taken in these cases. This
CL adds support, simply by passing the kAllowLargeObjectAllocation
argument to AllocateJSArray.

It also changes the fast path for strings to use this argument
rather than bailout to the runtime.

Bug: v8:7980, v8:8410
Change-Id: I18e88cb4ceb7ebeca250edd8b8b0eb401fdbd6e4