[array] Weaken bounds checks in Array.p.sort
If the array length is modified by user code during sorting, all bets
are off. We can only make assertions as compared to the original array
length.

Note that even though the algorithm may read and write from beyond
the current array length value, this adheres to the spec, which only
mandates that accesses are within the original array dimensions (i.e.: 0
<= i < original array length).

Bug: chromium:901633
Change-Id: Id765e80d4231ff6f2a73e040ec94c2b07f8c5b0f