Update patch set 4
Patch Set 4:

> Patch Set 3:
> 
> There's one more issue I forgot to mention in the bug. It's Chromium builds (as opposed to Google Chrome) that Linux distros include. Most of them use 'system ICU' instead of bundled ICU. 
> 
> We can take the same stance toward them as toward Node.js with system ICU. 
>    "To have Chromium 72, you need ICU 63."  
> 
> FYI, Blink does have similar conditionals (using ICU MAJOR VERSION macro and USING_SYSTEM_ICU) in a couple of places (Emoji and other character properties,  character encoding converters). 
> 
> In this CL, I don't think it's problematic to drop  a part where locale canonicalization is done using regex. The difference is not substantial and v8 would still compile. 
> 
> RelativeTimeFormat (formatToPart) change is a bit worrisome because it uses a new API in 63.1. So, it'll not compile with older ICU and we may as well be 'nice' by keeping the old work-around and conditionals.

Not RelativeTimeFormat. RelativeTimeFormat only use an additional enum.
ListFormat (formatToPart) use a total new API in 63.1 which is not part of this change at all but landed earlier already. Because when I submit that cl, we already require ICU 63 to compile there and I didn't include code to allow it be build on ICU prior 63. We have no trybot to ensure our code submission compile on earlier version of ICU so I don't believe our code, regardless w/ this cl or not, are ASSURED to be compiled on earlier version of ICU. If we DO want to support building on earlier version of ICU (which we currently do not since we have no trybot to assure them), then we should decide how many version we should support back and add trybots to ensure that. Without that, these #ifdef are simply "false sense of security" that make us "feel good" but complicated the code and in the mean time does not really support older versions.

Patch-set: 4
Reviewer: Gerrit User 1291594 <1291594@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>