[builtins] Support embedded builtins in nosnapshot builds
This CL adds support for embedded builtins in nosnap builds by creating
and setting an 'embedded blob' after builtin generation. Unlike
snapshot builds, the blob is not embedded into the .text section but
located on the C++ heap.

This makes nosnap builds more consistent with mksnapshot, and allows us
to simplify there and in serializer cctests.

Complications arise from the different workflows we need to support:
1. the standard mksnapshot build process,
2. serializer cctests that may create a fresh heap even in snapshot
   builds,
3. an inspector test that first creates a snapshot and later
   deserializes from it *in the same process*,
4. and nosnap builds.

Bug: v8:6666, v8:8350
Change-Id: I3842e40cdaf45d2cadd05c6eb1ec2f5e3d83568d