[esnext] implement private methods
This patch is a initial implementation of the stage 3 proposal
https://tc39.github.io/proposal-private-methods

- Implement the semantics of private methods/accessors

The design doc is in
https://docs.google.com/document/d/1T-Ql6HOIH2U_8YjWkwK2rTfywwb7b3Qe8d3jkz72KwA/edit?usp=sharing

The early errors for duplicate getter/setters do not work yet,
currently only one of them can be defined on the same class.
It should be fixable later with a local HashMap that checks for
duplicate accessors after the properties are parsed.

Bug: v8:8330
Change-Id: I55f0ebdf870f7807315a3ce8e3e7e38713233c4d