[parser] use ParsePropertyInfo as out parameter in object/class literals
This patch refactors the parsing of object literal properties and
class literal properties, putting the out parameters into a pointer of
struct `ParsePropertyInfo`. This struct is also aware of its potition
so `ParsePropertyName()` can also use this information to error
when parsing a private name in an object literal. It also makes
sure that the `ClassLiteralProperty::Kind` are all inferred
from the `ParsePropertyKind` and get used right away instead of
being passed around as out parameters.

Bug: 8330
Change-Id: I4c52592dfcaa3c8df30c4aba4c46e5c675acb394