Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > Patch Set 1:
> > 
> > This would cause regressions in relocation info size as well as code size on ARM and ARM64. We specifically made those two modes shareable for that reason. Here is an old measurement for {WASM_CALL}:
> > 
> > https://chromium-review.googlesource.com/c/v8/v8/+/1105759/4#message-019b0f9f92aeb8b99ec592f3adb0de6abbd42569
> > 
> > I don't have measurements for {WASM_STUB_CALL} at hand, but I remember the impact being even more visible for that mode. Here is the CL for this one:
> > 
> > https://chromium-review.googlesource.com/c/v8/v8/+/1091055/
> > 
> > Would it be possible to make this change PPC-only, to not affect other architectures? Also note that the shared entries that get relocated multiple times will be relocated to the same address (i.e. we never make entries have different targets that originally had the same target). Would it be possible to make the reloc iterator use this fact to still allow sharing?
> 
> Hello Michael,
> 
> Thanks for replying. Since ARM also uses constant pool, I think at some point it will suffer the same issue unless there is already mechanism to prevent the same entry get relocated twice. Yes, the shared entries should be pointing to the same target address and it's possible to make those shareable. The issue comes from v8/src/wasm/wasm-code-manager.cc line 568 and line 574. The set_wasm_call_address and set_wasm_stub_call_address could potentially reloc to the same constant pool entries (different PCs points to the same entry). Any suggestion on how to resolve this?

I think on ARM we don't actually emit a reloc-info for subsequent uses of a shared constant pool entry. The reloc info is recorded as part of {Assembler::ConstantPoolAddEntry} and omitted in case the entry is shared. This would also safe reloc-info size, on top of the constant pool size. I think we are doing something 

https://cs.chromium.org/chromium/src/v8/src/arm/assembler-arm.cc?type=cs&q=Assembler::ConstantPoolAddEntry&sq=package:chromium&g=0&l=5155

Could we do something similar for PPC? Another alternative would be to go with disabling the sharing as you intended with this CL, but make it specific to PPC. It would be totally fine to exclude the Wasm reloc modes in PPC's {Assembler::ConstantPoolAddEntry}, this would not affect other architectures and be a temporary solution for PPC. Ideally we would leave a TODO in the PPC code so that we eventually optimize the constant-pool and reloc-info size.

Patch-set: 1
Reviewer: Gerrit User 1118575 <1118575@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>