[scopes] Use MoveHead O(1) rather than ReinitializedHead O(n)
Since there are likely many unresolved references that remain, it's not a good
idea to walk the entire list everything time just to allow any random element
to become the head. We just need a member of the list to become the head. We
still need to find the member though, since the previous pointer to the member
was head_, which will contain a new element.

Change-Id: I65cd5752b432223d95cd529452a064d8dcc812e1