Remove unused paths in Runtime Load/StoreIC_Miss functions
RuntimeLoad/StoreIC_Miss are only called for the named properties. It
is not required to check for keyed and global kind in these functions.

Runtime_StoreIC_Miss is only used by the StoreIC which is used by
the bytecode handler for NamedProperties and by Turbofan when lowering
NamedProperties.

Runtime_LoadIC_Miss is only used by the LoadIC_BytecodeHandler, LoadIC
and LoadIC_Megamorphic. LoadIC_BytecodeHandler handles only named
properties and LoadIC and LoadIC_Megamorphic are used by Turbofan only
when lowering NamedProperties.

Bug:v8:8394
Change-Id: I8f55655af0dc727300eb78bfbd378d0011627a1c