Use --arm-arch for mksnapshot.
This removes the CAN_USE_* macros. With the exception of one flag for
MIPS, which I've left as it is, these macros were only used for Arm.

The build system now uses the existing --arm-arch option to mksnapshot
to select a code-generation target. The target used depends on the
existing arm_* flags.

This patch slightly changes the behaviour of --arm-arch, so that if a
specific target is given, it is used. Before, it would be silently
ignored if the host didn't support what was specified. The default value
is now "best", meaning that the dynamically-detected features are used.

This patch also removes some deprecated command-line options and their
associated support code.

Change-Id: I7f3caf3995349c3f57d187e3ecf10e10464f93de