[WIP] Multi-target snapshot support.
This is a proof of concept. The principle is that we run mksnapshot for
each --arm-arch target that we want to support, and use a mechanism
similar to that used for the "Trusted" embedded variant to select the
most appropriate snapshot based on the CPU features detected when V8 is
actually run.

We pay a cost in static code size, since the whole snapshot is
duplicated for each snapshot target, but since these are embedded
builtins, the physical memory cost does not increase with multiple V8
processes.

The infrastructure works for multi-target
snapshots on Arm, but some work remains. In particular:

- Verify that native builds work correctly.
- Verify that unrelated build configurations (such as other targets) are
  unaffected.
- Consider supporting non-embedded snapshots, if they are required.
- Make the target list configurable, and if necessary, refine the
  default set. Currently, the default set includes every --arm-arch
  setting at or above the C++ compiler target.
- The mechanism isn't really Arm-specific. The new Arm-specific paths
  should probably be replaced with some other flag.

Change-Id: Id87c605b06d192706759bbc7c0c55459b3070572