Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> (1 comment)
> 
> > Patch Set 1:
> > 
> > Thanks for the prototype!
> > 
> > My top-level feedback is that even if this feature causes even a modest executable size regression, there will be strong resistance. IIRC, it would add 1M+ to the Chrome ARM executable, and unless there is a substantial performance improvement for with the chip-specific optimizations, this is going to be a tough sell. How much of a performance difference do these make?

Yes, it's expensive for binary size, but cheap for dynamic memory usage. Of the ideas I'd considered, this seemed the most appropriate, but we will need to refine the default snapshot target list for each baseline to make sure we get the best cost/benefit balance. Instinctively, I'd say that shipping all four targets for a default Armv6 build is probably not worthwhile.

> > Is there any way that this mechanism can be done only per-function? It seems like there might be a sweet spot here, e.g. for the 10 (or X) performance-hottest builtins, generate platform-specific version of just those, which minimizes the additional on-disk footprint but would maximize performance benefit. Not sure if this is possible without adding some level of indirection that would also be costly.
> > 
> > Adding Jakob Gruber on CC, since we will certainly have thoughts about this given his work on embedded builtins and JIT-less V8.
> 
> 
> I would also be very interested in performance numbers.
> 
> I think implementing a mechanism to choose per function is going to be a lot of work, and very targeted towards certain benchmarks. You would be facing the size vs performance trade-off for every single builtin, too, and it doesn't sound sustainable to push back every time someone wants to improve a microbenchmark for *their* phone's CPU.

Per-function targeting is very appealing in principle because it means that we don't have to duplicate builtins that don't make use of the new features. However, as you say, it will be very complicated to implement (and maintain), and it's going to be hard to avoid some level of indirection. It is also constrained by the range of direct branches.

I expected the approach I've used here to be useful for the features which we'll want to use everywhere. For example, Armv7 movw/movt gets used very frequently when available, and this would be a good approach for some Arm64 security features which change the prologue and epilogue of every function.

Thanks for the review,
Jacob

Patch-set: 1
Reviewer: Gerrit User 1119620 <1119620@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>