Update patch set 2
Patch Set 2:

> Patch Set 1:
> 
> > Patch Set 1:
> > 
> > (1 comment)
> > 
> > Is this still needed? Shouldn't this work by now?
> 
> It doesn't occur anymore due to arm32 supporting f32 and so the case not coming up, but this is still necessary to cover the case as has been done for CallDirect.

Originally the idea was that the Liftoff compiler should work independent of the Liftoff assembler. So if the assembler does not support something, it can bail out (and thus mark the result invalid), but this should have no influence on the further compilation. So it should be ok to just check for assembler bailouts in the end. 
For that reason, I would prefer to skip this CL and even revert the previously introduced check now that we don't need it any more.

Patch-set: 2
Reviewer: Gerrit User 1177599 <1177599@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>