Update patch set 1
Patch Set 1:

(1 comment)

One hand I like the idea of not having a runtime function for this but on the other hand, it's nice to check the default locale. This test is a bit too loose

Can you set the default locale using environment vars and then check it? see mjsunit/icu-date-to-string.js test for an example

Patch-set: 1
Reviewer: Gerrit User 1158954 <1158954@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>