Update patch set 6
Patch Set 6:

> Patch Set 4:
> 
> > Patch Set 4:
> > 
> > I think we need to take a step back and take a look at the various validations that are happening already. We have validation logic in JSCollator::Initialize, JSDateTimeFormat::Intialize, etc .. should those be moved to LookAndValidateUnicodeExtensions The collation validation implemented in this CL is duplicated in JSCollator::Initialize, right? 
> > 
> > Can we come up with a nice design for how to do all the validations holistically?
> 
> To fix intl402/((DateTime|Number)Format|Collator)/ignore-invalid-unicode-ext-values we should validate inside ResolveLocale because step 8.h.ii.1.a. It is true that we have some other place already have validation logic and we should remove those.
> 
> Now sure how better "a nice design for how to do all the validations holistically" need to be. 
> Based on "step 8.h.ii.1.a" the check is "If keyLocaleData contains requestedValue," but I don't believe there is a common API for such "contains" check in ICU across all the api.

By placing all the logic in LookupAndValidateUnicodeExtensions we're doing unnecessary validations. There's no need to do a validation for collation when we create a number format object. This is what I mean by a better design that makes sure that we do only the required validations. Should we pass in a filter lambda function that does the validation? Should we pass in a list of valid key-value pairs to do the validation?

Patch-set: 6
Reviewer: Gerrit User 1158954 <1158954@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>