Update patch set 23
Patch Set 23:

> Patch Set 23:
> 
> > Patch Set 23:
> > 
> > I hijacked your CL here https://crrev.com/c/1362049 and will see if I find something obvious that blocks the mac bot...
> 
> There seem to be some particular test262 tests hanging on mac. When tests are hanging on mac, they sometime manage to block the whole machine in a way that we don't get stdout flushed. I experimented a bit and flushed it manually and the test runner seems to start running tests and then hangs on some. The usual threads that keep running in the background and keep the runner alive also seem to not work on Mac, when tests hang in this way.
> 
> I suggest you try to repro locally on a Mac. You can also try to trigger the trybot and only run a portion of the tests if you have a suspicion. I e.g. just triggered this:
> git cl try -b v8_mac64_rel_ng -B luci.v8.try -p 'testfilter=["language/statements/class/private-accessor-name/*", "language/expressions/class/syntax/valid/grammar-privatemeth-duplicate-get-set", "language/statements/class/syntax/valid/grammar-privatemeth-duplicate-get-set"]'
> 
> Those are some of the tests marked now in the status file. Maybe you can trigger only on those tests that are added now? I didn't know how to actually get to them.

Does the command above specifically run those tests? Those are the tests marked as FAIL in this patch because the features being tested are not implemented correctly yet, and they currently lead to normal errors. Locally when I run

$ tools/run-tests.py --outdir out.gn/x64.release test262/language/statements/class/*
$ tools/run-tests.py --outdir out.gn/x64.release test262/language/expressions/class/*

on a mac with them marked as FAIL, it does not hang. And if I comment out the entries in the status file, I get the failures as expected but it does not hang either. Should I mark them as SKIP for this patch and see if it works for the buildbots?

Patch-set: 23
Reviewer: Gerrit User 1288325 <1288325@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Commit-Queue=0