Update patch set 1
Patch Set 1:

> Patch Set 1: Code-Review+1
> 
> LGTM, but could we add a serializer test (like this one https://cs.chromium.org/chromium/src/v8/test/cctest/test-serialize.cc?q=test-serialize&sq=package:chromium&g=0&l=912) for strings? I want to make sure that we do not require embedders who use custom startup snapshot to use internalized strings. Iirc regular strings are not put onto RO-space, so we do not have to require them to be internalized?

I'm not sure how to write this test. If I write:

function f() { return "foo"; }

then "foo" will be internalized because it's a literal in the source code. I could construct the string by writing "foo" + "bar", then it won't be internalized but also won't be placed in the snapshot unless the string is constant folded but now then the test would be quite fragile as it's relying on behaviour that's nothing to do with what we're really trying to test.

Patch-set: 1
Reviewer: Gerrit User 1204345 <1204345@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Commit-Queue=0