Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> (9 comments)
> 
> Some thoughts below...I can't say this immediately strikes me as a big win. And as I note, there's at least one refactor here (ParseSingleExpression) that could be done today, with ExpressionClassifier, and that makes it harder for me to evaluate how much simpler ExpressionScope makes things, vs just removing code duplication around the use of either the Classifer or Scope.
> 
> Your only mention of "performance" in the CL description says this doesn't affect perf. Do you hope that some follow-ons from this would give perf gains in future? Or is this CL mostly about style/complexity/understandability/correctness?

Right now the performance win isn't that big yet, although it does seem to be slightly better on my machine. The idea is that this makes it easy to test at runtime why we're parsing expressions; and as I said, I'd like to introduce multiple scope types to distinguish them further and have smaller scope objects. E.g., something like DeclarationParsingScope would not need to track errors at all I think, because we know what we're parsing; similar to what we already do for "let". The scopes should just be a way to configure the parser, and allow us to store information on the side if we don't know yet.

Once we know exactly where we are parsing expressions that are actually declarations, we can immediately declare them rather than track them through the expressions. In the case of arrow functions we can use a ScopedPtrList to track variable names that may want to turn into declarations. That gets rid of the list pointer in PreParserExpression and of the ThreadedList through the VariableProxies. That will improve performance. Implementing that is much simpler with this version since it directly aligns.

What I like already is that the scopes better force correctness than before; guaranteeing that they are validated. What I don't like so much is the duplication that I added for identifier parsing, but if the scopes are lightweight enough I should be able to reintroduce that later.

Patch-set: 2
Reviewer: Gerrit User 1184114 <1184114@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Commit-Queue=0