Update patch set 24
Patch Set 24:

We are currently creating the closures in the initializer function so the JSFunction associated  with the same method in each instance are different. However, we cannot create the closure during class evaluation because the context would be different. Maybe we can implement this by creating the closure during the first call of the initializer function, store it onto the context, and reuse that in subsequent calls? (I believe the closures would not be in immutatble context slots anymore, though)

Patch-set: 24
Reviewer: Gerrit User 1288325 <1288325@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Commit-Queue=0