Update patch set 24
Patch Set 24:

> Patch Set 24:
> 
> We are currently creating the closures in the initializer function so the JSFunction associated  with the same method in each instance are different. However, we cannot create the closure during class evaluation because the context would be different. Maybe we can implement this by creating the closure during the first call of the initializer function, store it onto the context, and reuse that in subsequent calls? (I believe the closures would not be in immutatble context slots anymore, though)

Hmm, do you know what was wrong with your previous analysis?

To avoid creating multiple JSFunctions, can you create the closure when creating the class, and stash the private methods on the prototype, keyed on the same private symbol used for the instances?

Stepping back, I think we don't need to depend on any sort of complex optimization to get good space usage, and instead organize the data structures and brand checks more directly. I am working on an editorial change to the private methods spec to make it more clear how that could be organized.

Patch-set: 24